refactor: extract create sensor use case and ms-temperature-monitoring to clean architecture

+++ microservices/temperature-monitoring/src/main/java/com/algasensors/temperature/monitoring/api/model/SensorAlertInput.java
@@ -2,8 +2,10 @@ package com.algasensors.temperature.monitoring.api.model;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class SensorAlertInput {
-    private Double maxTemperature;
-    private Double minTemperature;
+    private BigDecimal maxTemperature;
+    private BigDecimal minTemperature;
 }
