+++ microservices/temperature-monitoring/src/main/java/com/algasensors/temperature/monitoring/api/model/TemperatureLogData.java
@@ -6,8 +6,8 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.OffsetDateTime;
-import java.util.UUID;
 
 @Data
 @Builder
@@ -17,5 +17,5 @@ public class TemperatureLogData {
 private String id;
 private TSID sensorId;
 private OffsetDateTime registeredAt;
-private Double value;
+private BigDecimal value;
 }
