+++ microservices/temperature-monitoring/src/main/java/com/algasensors/temperature/monitoring/domain/model/TemperatureLog.java
@@ -7,6 +7,7 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.OffsetDateTime;
 
 @Data
@@ -21,7 +22,7 @@ public class TemperatureLog {
     private TemperatureLogId id;
 
     @Column(name = "temperature_value")
-    private Double temperatureValue;
+    private BigDecimal temperatureValue;
 
     private OffsetDateTime registeredAt;
